+++ packages/BilibiliRecorder/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 1.6.0
 
-- 支持 `recorderType` 参数用于配置底层录制器，支持`ffmpeg | mesio`
+- 支持 `recorderType` 参数用于配置底层录制器，支持`auto | ffmpeg | mesio`
 - 修复 `videoFormat=auto` 时某些情况下格式的判断
 -